feat: notify controllers when outputs are ready to be destroyed
Controllers should use two-phased approach when destroying their
outputs:

* first, resource is marked as being torn down (`Teardown()`), this also
prevents adding new finalizers
* second, when finalizers are removed by dependent controllers,
controller can finally completely remove its out with `Destroy()`

This problem is that controller should wait for the finalizers to be
empty between phases one and two. Waiting is possible, but controller
might not want to block other reconciliation activities while waiting for
the output to be ready to be destroyed.

This PR addresses this issue by allowing to declare controller output as
an input of special kind `InputDestroyReady` which means controller is
only interested to be notified when output is ready to be destroyed.

Signed-off-by: Andrey Smirnov <smirnov.andrey@gmail.com>