feat: implement persistent backing store for resources
Resources are still stored and processed completely in-memory, but
backing store is updated synchronously with each resource update and
delete. When COSI state is created, contents of the namespace will be
loaded from the persistent store on first access.

There is a first implementation of backing store implemented using
BoltDB (etcd `bbolt` fork).

Persistence requires a way to marshal each resource to bytes and back,
there is a first implementation provided using resource protobuf
serialization.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>