feat: add support for no protobuf unmarshaling via gRPC
This allows to skip full protobuf unmarshaling and preserve the original
YAML representation which is sometimes helpful when there's a goal to
preserve compatibility with the server when proto definitions are not in
sync.

E.g. getting only YAML representation of a resource which is not known
to the client, or some fields are not supported by the client yet.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>