fix: always destroy tearing down outputs
The problem is triggered with the following flow: let's image transform
controller maps resources A to resources B keeping ID the same.

1. A(1) is created.
2. Transform controller creates output B(1).
3. Something sets a finalizer on B(1).
4. A(1) is destroyed.
5. Transform controller can't destroy B(1), as it has a finalizer set,
   but it puts B(1) into tearing down phase.
6. A(1) is re-created (that is essentially new A(1)).
7. Transform controller can't update B(1), as it's being torn down.
8. Finalizer from B(1) is removed.
9. Transform controller still can't move on.

Before the fix controller will be "stuck" at the step (9), while with
the fix controller will first destroy tearing down B(1), and then
re-create it from new A(1).

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>