feat: allow transform controller to ignore tearing down inputs
This is sometimes useful for `Resource` -> `ResourceStatus` when the
status resource should exist as long as the input resource exists, even
if it's being torn down.

This only works if the controller doesn't set finalizers on its input,
as otherwise it breaks the logic: if the finalizer is set, output should
be destroyed before the input.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>