feat: add new package testutils
Since `*testing.T` is explicitly not goroutine safe (and there is even `go vet` check), you can't create new goroutines in
test and use `*testing.T`. One approach is to return errors explicitly from tests helpers, but this doesn't scale well. This
commit does another - it provides a safe wrapper around any implementation of `*testing.T` that accepts and returns a context,
which will be canceled on any test failure.

As an example - it fixes a small concurrent bug in `setup` function in `transform` package tests.

Signed-off-by: Dmitriy Matrenichev <dmitry.matrenichev@siderolabs.com>