feat: provide an example of typed resource type
This is something that I wanted to try for a while - attach the resource type directly to the resource.Type constant.
The (partial) result is this PR. While it may look interesting, I don't think that we should merge it ATM.

Because:
1. There are a LOT of `Naked()` calls around the code.
2. This will require a lot of refactoring. Even here it required a lot of places to be changed.
3. I'm not sure that this level of generic abuse is worth it (will it increase the resulting binary size because of the bloat?).

Still, this is the good starting point if anyone is interested in how it could be done.

Signed-off-by: Dmitriy Matrenichev <dmitry.matrenichev@siderolabs.com>