fix: use destroy ready input type in CleanupController
This should reduce a lot number of wakeups of the `CleanupController`
when it shouldn't actually do any work.

As the controller in RemoveOutputs mode always tearsdown and destroy all
outputs owned by configured owner, it is only interested in the moment
the output has no finalizers left.

Improve the tests to actually verify the cleanup with finalizers.

Fix typos.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>