fix: push to queue with non-zero timestamp
If we always push with zero timestamp, there's a chance that items with
non-zero timestamp (e.g. items pushed with some backoff) will never be
processed at all, as they will stay at the tail of the queue, and the
head of the queue will be occupied with items with zero timestamp.

With this change, new items without backoff are pushed with timestamp
`now`, so they will be properly ordered with other items.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>