feat: allow setting a custom logger for the user operations
Allow passing a `UserLogger` to the runtime, which will be used where the control is passed to the user of the runtime, i.e., the implementor of the controllers.

With this, it is possible to have different log levels/configuration between the runtime internals and the application code that is using it as its framework.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>