fix: validate QController mapped results
There were no checks on what map function returns as a result, which
might lead to fun bugs when map function returns an input which is not
an expected intput of the QController, and it's hard to track why it got
into the queue.

This validation isn't strictly required by the runtime itself, but it
might lead to hard to trace bugs for the users of the runtime.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>