feat: provide an option to skip cache on reads in the controller
Usually all reads are cached, and that cache is consistent with the
watch stream and controller notifications. The cache is eventually
consistent (it will catch up and notification will be delivered).

But sometimes controllers make decisions which are not reconcilable back
when the cache catches up. One of the case was in Omni with automatic
scaling: the controller might try to scale up the same machine set
twice, and it might succeed to do so, but at the same time there's no
good way sometimes to "reconcile back", as scale down is not symmetric
with scale up (additional pre-checks).

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>