feat: implement correct watch restart for controller runtime
Controller runtime does `WatchKind` on all controller input resources,
but initial bookmark value is only delivered if there is any update
received (on each of the watches).

If resource cache is used, `BootstrapContents` will deliver the
bookmark, but if the cache is not used, and there are no updates, no
bookmark is ever sent, so watch can't be restarted.

Fix this by adding yet another option to send 'Noop' events which carry
only the bookmark allowing watch retry code to handle it.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>