fix: correct deduplication of watch events
This is extracted from #633 as a separate fix.

The idea is that if we merge/deduplicate watch events, we should do this
by immutable attributes of the resource: namespace, type, id.

The previous implementation might split two events on e.g. finalizers
being empty, so two events which report finalizers empty/finalizers
non-empty might be delivered in the wrong order: first empty, next
non-empty.

It might not trigger a real bug due to the way controllers work, but
still is not consistent and correct.

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>